// src/app/shared/models/enums/reservation-request-state.enum.ts
export enum ReservationRequestState {
  CONFIRM_AWAITING = 'CONFIRM_AWAITING',
  NOT_ALLOCATED = 'NOT_ALLOCATED',
  ALLOCATED = 'ALLOCATED',
  ALLOCATED_STARTED = 'ALLOCATED_STARTED',
  ALLOCATED_STARTED_NOT_AVAILABLE = 'ALLOCATED_STARTED_NOT_AVAILABLE',
  ALLOCATED_STARTED_AVAILABLE = 'ALLOCATED_STARTED_AVAILABLE',
  ALLOCATED_FINISHED = 'ALLOCATED_FINISHED',
  FAILED = 'FAILED',
  DENIED = 'DENIED',
  MODIFICATION_FAILED = 'MODIFICATION_FAILED',
}